pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dependency-injector>=4.48.2",
    "fastapi>=0.117.1",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.36.0",
]
